requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
prompt_toolkit==3.0.52
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rabbitmq==0.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
